Right justify numerical tabular html output (Sébastien Aperghis-Tramoni). Work around ExtUtils::MakeMaker realclean bug (Florian Ragwitz) (rt 17324). Don't try to delete db if it isn't there.

+++ lib/Devel/Cover/DB.pm
@@ -107,7 +107,7 @@ sub delete
     opendir DIR, $db or die "Can't opendir $db: $!";
     my @files = map "$db/$_", map /(.*)/ && $1, grep !/^\.\.?/, readdir DIR;
     closedir DIR or die "Can't closedir $db/runs: $!";
-    rmtree(\@files);
+    rmtree(\@files) if @files;
     $self
 }
